# openapi.yaml
openapi: 3.1.0

info:
  title: task-api
  version: 1.0.0
  summary: タスクの作成・取得・更新・削除ができるREST API
  description: |
    Node.js + TypeScript でのAPI設計、OpenAPI仕様書の作成、コンテナ化を実践するための学習用プロジェクト。

    ## データの永続性
    ユーザーとタスクは PostgreSQL に保存する。API サーバーを再起動してもデータは失われない。

    ## タスクの所有者
    タスクは作成したユーザーのものになり、本人しか取得・更新・削除できない。
    他人のタスクの ID を指定した場合は、存在しない ID と同じ 404 を返す
    (403 を返すと、その ID のタスクが存在することを他人に知らせてしまうため。RFC 9110 §15.5.4 が認める扱い)。

    ## エラー応答
    すべてのエラーは RFC 9457 (Problem Details for HTTP APIs) 形式で返し、
    `Content-Type: application/problem+json` を付与する。

    ## 認証
    `/tasks` 配下のすべての操作は、`POST /auth/login` で取得した JWT を
    `Authorization: Bearer <accessToken>` ヘッダで送る必要がある。
    トークンの有効期限は発行から10分。期限が切れたら、再度ログインしてトークンを取り直す。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000
    description: ローカル開発環境

# 既定ではすべての操作に認証を要求する。認証が不要な操作は、個別に `security: []` で上書きする。
security:
  - bearerAuth: []

tags:
  - name: auth
    description: ユーザー登録とログイン(トークンの発行)
  - name: tasks
    description: タスクの管理
  - name: health
    description: サービスの稼働確認

paths:
  /health:
    get:
      tags: [health]
      summary: 稼働確認
      description: サービスが応答可能かを返す。ロードバランサのヘルスチェック用。
      operationId: getHealth
      security: []
      responses:
        "200":
          description: 稼働中
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok

  /auth/register:
    post:
      tags: [auth]
      summary: ユーザーを登録する
      description: |
        ユーザーを新規作成する。パスワードは scrypt でハッシュ化して保存し、レスポンスには含めない。
        登録してもトークンは発行されないため、続けて `POST /auth/login` を呼ぶ。

        ユーザー名の重複は DB の一意制約で判定する。同じユーザー名の登録が同時に届いても、
        成功するのは1件だけで、残りは 409 になる。
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"

  /auth/login:
    post:
      tags: [auth]
      summary: ログインしてアクセストークンを取得する
      description: |
        ユーザー名とパスワードが正しければ、アクセストークン(JWT)を発行する。

        ユーザー名が存在しない場合とパスワードが違う場合は、区別せずに同じ 401 を返す
        (応答時間も揃えている)。どちらが誤っているかを外部から判別できないようにするため。
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: ユーザー名またはパスワードが違う
          headers:
            WWW-Authenticate:
              schema:
                type: string
              example: Bearer
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: about:blank
                title: Unauthorized
                status: 401
                detail: ユーザー名またはパスワードが違います。
                instance: /auth/login

  /tasks:
    get:
      tags: [tasks]
      summary: タスク一覧を取得する
      description: ログイン中のユーザーが所有するタスクを、作成日時の昇順で返す。1件もなければ空の配列を返す。
      operationId: listTasks
      responses:
        "200":
          description: タスクの配列
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [tasks]
      summary: タスクを作成する
      description: |
        タスクを新規作成する。所有者はログイン中のユーザーになる。
        `id` / `createdAt` / `updatedAt` はサーバーが採番するため、
        リクエストに含めてはならない。未知のプロパティを含むリクエストは 400 で拒否する。
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: 作成成功
          headers:
            Location:
              description: 作成されたタスクのURI
              required: true
              schema:
                type: string
              example: /tasks/f4b28c3f-1844-4619-b70e-b3cc26267617
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /tasks/{id}:
    parameters:
      - $ref: "#/components/parameters/TaskId"

    get:
      tags: [tasks]
      summary: タスクを1件取得する
      operationId: getTask
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"

    patch:
      tags: [tasks]
      summary: タスクを部分更新する
      description: |
        指定したプロパティのみを更新する。省略したプロパティは変更されない。
        更新が成功すると `updatedAt` がサーバーによって更新される。

        リクエストボディが空オブジェクト `{}` の場合は 400 を返す。
        意図しない空更新を検出するため。
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"

    delete:
      tags: [tasks]
      summary: タスクを削除する
      description: |
        指定したタスクを削除する。

        ## 再送時の扱い

        DELETE は冪等 (RFC 9110 §9.2.2) であり、通信断などで応答を受け取れなかった場合、
        クライアントは安全に再送できる。ただし本APIは既に存在しないタスクへの DELETE に 404 を返すため、
        **再送によって 404 を受け取った場合、クライアントは削除が成功したものとみなしてよい**。

        ## 404 を返す設計にした理由

        常に 204 を返す設計と比較検討した上で 404 を採用している。

        1. 常に 204 にすると、クライアントは「自分の操作で実際に削除が起きたのか、
           元から存在しなかったのか」を知る手段を失う。404 を返しておけばクライアント側で
           成功として扱えるが、逆方向の復元はできない
        2. `GET /tasks/{id}` が 404 を返すため、同一URLのリソース存在性について一貫した回答になる
      operationId: deleteTask
      responses:
        "204":
          description: 削除成功。レスポンスボディは空。
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `POST /auth/login` で取得したアクセストークン。
        Swagger UI では「Authorize」ボタンに、`Bearer ` を付けずにトークンだけを入力する。

  parameters:
    TaskId:
      name: id
      in: path
      required: true
      description: |
        タスクのID。UUID の形式でない値や、他人のタスクの ID を指定した場合は 404 を返す。
      schema:
        type: string
        format: uuid
      example: f4b28c3f-1844-4619-b70e-b3cc26267617

  schemas:
    Credentials:
      type: object
      required: [username, password]
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: "^[a-zA-Z0-9_-]+$"
          description: 英数字と `_` `-` のみ
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
      example:
        username: alice
        password: password123

    User:
      type: object
      required: [id, username, createdAt]
      properties:
        id:
          type: string
          format: uuid
          description: サーバーが採番するユーザーID。トークンの `sub` にも入る。
        username:
          type: string
        createdAt:
          type: string
          format: date-time
      example:
        id: 6616068c-14aa-4388-a3da-06481f354d59
        username: alice
        createdAt: "2026-09-27T07:30:00.000Z"

    TokenResponse:
      type: object
      required: [accessToken, tokenType, expiresIn]
      properties:
        accessToken:
          type: string
          description: "JWT(HS256)。`Authorization: Bearer <accessToken>` で送る。"
        tokenType:
          type: string
          const: Bearer
        expiresIn:
          type: integer
          description: トークンが失効するまでの秒数
          example: 600

    TaskStatus:
      type: string
      enum: [todo, doing, done]
      description: |
        タスクの進捗状態。
        - `todo`: 未着手
        - `doing`: 着手中
        - `done`: 完了

    Task:
      type: object
      required: [id, title, description, status, dueDate, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          description: サーバーが採番するタスクID
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: [string, "null"]
          maxLength: 1000
        status:
          $ref: "#/components/schemas/TaskStatus"
        dueDate:
          type: [string, "null"]
          format: date-time
          description: 期限。RFC 3339 形式。
        createdAt:
          type: string
          format: date-time
          description: 作成日時。サーバーが採番する。
        updatedAt:
          type: string
          format: date-time
          description: 最終更新日時。サーバーが採番する。
      example:
        id: f4b28c3f-1844-4619-b70e-b3cc26267617
        title: OpenAPI仕様書を書く
        description: null
        status: todo
        dueDate: "2026-09-30T12:00:00Z"
        createdAt: "2026-09-21T07:27:52.668Z"
        updatedAt: "2026-09-21T07:27:52.668Z"

    CreateTaskRequest:
      type: object
      required: [title]
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: [string, "null"]
          maxLength: 1000
          default: null
        status:
          $ref: "#/components/schemas/TaskStatus"
          default: todo
        dueDate:
          type: [string, "null"]
          format: date-time
          default: null
      example:
        title: OpenAPI仕様書を書く
        dueDate: "2026-09-30T12:00:00Z"

    UpdateTaskRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      description: 更新したいプロパティを1つ以上指定する。空オブジェクトは 400 になる。
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: [string, "null"]
          maxLength: 1000
        status:
          $ref: "#/components/schemas/TaskStatus"
        dueDate:
          type: [string, "null"]
          format: date-time
      example:
        status: doing

    Problem:
      type: object
      description: RFC 9457 (Problem Details for HTTP APIs) 形式のエラー応答
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri
          description: 問題の種類を識別するURI。特定していない場合は `about:blank`。
          default: about:blank
        title:
          type: string
          description: 問題の種類を表す短い人間可読な要約
        status:
          type: integer
          description: HTTPステータスコード
        detail:
          type: string
          description: この発生事例に固有の説明
        instance:
          type: string
          description: 問題が発生したリソースのURI
      example:
        type: about:blank
        title: Not Found
        status: 404
        detail: ID 'f4b28c3f-1844-4619-b70e-b3cc26267617' のタスクは存在しません。
        instance: /tasks/f4b28c3f-1844-4619-b70e-b3cc26267617

    ValidationProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            errors:
              type: object
              description: バリデーション失敗の内訳 (RFC 9457 の拡張メンバー)
              properties:
                formErrors:
                  type: array
                  description: オブジェクト全体に対する誤り (未知のプロパティ、空オブジェクト等)
                  items:
                    type: string
                fieldErrors:
                  type: object
                  description: プロパティごとの誤り
                  additionalProperties:
                    type: array
                    items:
                      type: string
      example:
        type: about:blank
        title: Bad Request
        status: 400
        detail: リクエストボディが不正です。
        instance: /tasks
        errors:
          formErrors:
            - 'Unrecognized key: "titel"'
          fieldErrors:
            title:
              - "Invalid input: expected string, received undefined"

  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblem"

    NotFound:
      description: 対象のリソースが存在しない(他人のタスクを指定した場合も含む)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

    Unauthorized:
      description: |
        認証が必要、またはトークンが無効。`WWW-Authenticate` ヘッダで理由を区別する。
        - トークンが送られていない: `Bearer`
        - トークンが不正・改ざん・期限切れ: `Bearer error="invalid_token"`(再ログインしてトークンを取り直す)
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer error="invalid_token"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: about:blank
            title: Unauthorized
            status: 401
            detail: トークンが無効です。
            instance: /tasks

    Conflict:
      description: 現在のリソースの状態と衝突する(ユーザー名が既に使われている等)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: about:blank
            title: Conflict
            status: 409
            detail: ユーザー名は既に使用されています。
            instance: /auth/register